+++ src/main/java/roomescape/dto/ReservationDto.java
@@ -1,4 +1,4 @@
-package dto;
+package roomescape.dto;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
